+++ .github/workflows/CI-windows.yml
@@ -128,6 +128,7 @@ jobs:
           python -m pip install pip --upgrade || exit /b !errorlevel!
           python -m pip install pytest || exit /b !errorlevel!
           python -m pip install pytest-custom_exit_code || exit /b !errorlevel!
+          python -m pip install pytest-timeout || exit /b !errorlevel!
 
       - name: Run CMake
         if: false # TODO: enable
