+++ test/cli/test-other.py
@@ -3,6 +3,7 @@
 
 import os
 import pytest
+import time
 
 from testutils import cppcheck
 
@@ -141,3 +142,19 @@ def test_progress_j(tmpdir):
     assert exitcode == 0
     assert stdout == "Checking {} ...\n".format(test_file)
     assert stderr == ""
+
+
+@pytest.mark.timeout(10)
+def test_hang_array_many_strings(tmpdir):
+    # 11901
+    # cppcheck valueflow hangs when analyzing a file with many strings
+    filename = os.path.join(tmpdir, 'hang.c')
+    with open(filename, 'wt') as f:
+        f.write("const char *strings[] = {\n")
+        for i in range(20000):
+            f.write('    "abc",\n')
+        f.write("};\n")
+    start_time = time.time()
+    cppcheck([filename])
+    duration = time.time() - start_time
+    assert duration < 10
